refactor(manage-users): inline table page logic and drop shared ManagementTablePage
