+++ src/components/dashboard/table/Table.tsx
@@ -1,3 +1,4 @@
+import { Copy, TrendingUp } from "lucide-react";
 import type { FC, ReactElement, ReactNode } from "react";
 import {
   useCallback,
@@ -7,7 +8,6 @@ import {
   useRef,
   useState,
 } from "react";
-import { Copy, TrendingUp } from "lucide-react";
 import { AiOutlineDelete } from "react-icons/ai";
 import { HiOutlinePencil } from "react-icons/hi";
 import Loader from "@/app/loading";
